+++ Source/Core/SimplePolylineGeometry.js
@@ -10,7 +10,8 @@ define([
         './Geometry',
         './GeometryAttribute',
         './GeometryAttributes',
-        './Color'
+        './Color',
+        './Cartesian3'
     ], function(
         defined,
         DeveloperError,
@@ -22,7 +23,8 @@ define([
         Geometry,
         GeometryAttribute,
         GeometryAttributes,
-        Color) {
+        Color,
+        Cartesian3) {
     "use strict";
 
     /**
@@ -33,7 +35,7 @@ define([
      * @constructor
      *
      * @param {Array} options.positions An array of {@link Cartesian3} defining the positions in the polyline as a line strip.
-     * @param {Array} [options.colors] An Array of {@link Color} defining the per vertex or per segment colors.
+     * @param {Array} [options.colors=undefined] An Array of {@link Color} defining the per vertex or per segment colors.
      * @param {Boolean} [options.colorsPerVertex=false] A boolean that determines whether the colors will be flat across each segment of the line or interpolated across the vertices.
      *
      * @exception {DeveloperError} At least two positions are required.
@@ -100,9 +102,8 @@ define([
 
             var color;
             if (perSegmentColors && i > 0) {
-                positionValues[j++] = p.x;
-                positionValues[j++] = p.y;
-                positionValues[j++] = p.z;
+                Cartesian3.pack(p, positionValues, j);
+                j += 3;
 
                 color = colors[i - 1];
                 colorValues[k++] = Color.floatToByte(color.red);
@@ -115,9 +116,8 @@ define([
                 break;
             }
 
-            positionValues[j++] = p.x;
-            positionValues[j++] = p.y;
-            positionValues[j++] = p.z;
+            Cartesian3.pack(p, positionValues, j);
+            j += 3;
 
             if (defined(colors)) {
                 color = colors[i];